Reorganized employee dashboard
