+++ frontend/src/Pages/EmployeeDashboard/EmployeeDashboardComponents/Inventory.jsx
@@ -1,0 +1,7 @@
+
+/**
+ * 
+ */
+export default function Inventory() {
+
+}
