+++ frontend/src/Pages/EmployeeDashboard/EmployeeDashboardComponents/Orders.jsx
@@ -1,0 +1,7 @@
+
+/**
+ * 
+ */
+export default function Orders() {
+    
+}
